Adjusted blending style for extras.
This change fixes the blending settings for each part individually so that they allow the maximum amount of customization, rather than show themselves too white or too dark.
This doesn't allow FULL colour customization for some parts, for example the horns refuse to go full black and struggle with going a darker shade of any colour, but it's the best I could do.