Fixes #37 Unathi horns and frills displayed as ERROR on preview as well as Tajaran and Skrell specific hair styles also showing the same error. Fixes a Unathi, Tajaran and Vulpkanin tail colouration issue. Fixes #26 wings and tails from the extras menu not showing when facing north. (#133)
* #37 Fixed unathi horns/frills ERROR on preview.

ERROR display is fixed from the preview and Unathis can now have frills and horns again!

- Fixed unathi-related horns and frills in the extra.dm file.

- Created new unathi_extras.dmi file to organize all Unathi related extras.

- Added several additional horns and frills to the available options.

- unathi_extras.dmi contains several spines and tails that could potentially be added as optional extras if the character creator menu ever gets improved and has an "extras" option available.

- Fixed _mech_setup.dm and mech.dm so that they allow compiling. This is accurately changed according to another pull request upstream. I think...

* Set _mech back to how it was and disabled setup and undefines.

* Fixes #19 Body colouration code for Unathi.

It fixes the compiling error on windows machines by changing "_mech.dm" to "mech.dm" since BYOND sorts it alphabetically.

Fixes #19 issue regarding Unathi colouring issues. The Highlander Zorren colouration "issue" is not an issue, it is a lore-based colouration change. They tend to have a darker colour palette.

Adds Unathi horns to the Ears section on character customization extras. This allows having horns or spikes as well as frills.

* Adjusted blending style for extras.

This change fixes the blending settings for each part individually so that they allow the maximum amount of customization, rather than show themselves too white or too dark.
This doesn't allow FULL colour customization for some parts, for example the horns refuse to go full black and struggle with going a darker shade of any colour, but it's the best I could do.

* Commenting

Added comments to bring it up to standards.

* Fixes #26 and additional hair style ERRORs.

- #26 is fixed. Wings and tails now properly display when facing north.

- Tarajan and Vulpkanin tails no longer have a colouration issue and now display the same colour selected as the body colour.

- Tajaran and Skrell specific hair styles no longer display as an ERROR on the preview and work as intended in-game.

- Added breakers to human_face.dmi to identify the Tajaran specific hair styles quicker and easier, as it was not very clear at all before.

* Fixed typo

Fixed a typo where I accidently wrote 10 instead of 20.

* Fixes tail and wings overlay issue.

- Tails and wings from the Species Customization menu now hide properly when behind large chairs or bed covers.

- Held items now display over the tail and wings layers, as intended.