Adds a Copy Slot button #32
Adds a simple copy slot button to the character menu. It loads all the currently loaded slot's contents onto the slot you select, which can then be saved to that new slot.

Tested with changes on every single menu, copying over to both default and modified slots.

The only issue as of now is that the Load Slot dialogue box doesn't show the names of your saved characters, and instead defaults to the "Character[i]" statement.