Lasblender damage adjustment (#5144)
* Damage adjustment

This may actually just be too low, as we want players to specifically choose this thing, but with weapon attachments the damage output can be adjusted significantly. closes #5142

* Improves damage output

After some discussion, the damage multiplier was increased to make it, at least, not worse than the weapon it is replacing.
Conflicts:
	code/modules/projectiles/guns/energy/lasersmg.dm