Neotheo QOL Tweaks (#5435)
* Entreaty Update

Entreaty now calls out the location of the Cruciformed.

This was agreed to be a fair change last year, I'm not sure whether opinion would have changed on that.

* Tweaked Bounty Cost

Adjusts bounty cost to be more reasonable. As it stands, most Inquisitors just keep their uplink tab open rather than recasting it, due to the absurd power cost. As with all other changes in this PR, this was agreed upon as balanced a year ago.

Also minor grammar fix.