More staverbs updates (#5660)
* Update staverbs.dm

* Fixed oopsie

* UPDATE: Fixed jank

removed jank. interaction now takes longer, with the time greatly reduced by your COG stat. also adds a var to the computers to make it easier to track if they've been hacked, and notifies the user if the computer is hacked already when they try to interact with it.

* More staverb stuff
Conflicts:
	code/controllers/subsystems/staverbs.dm