Updates Excelsior mines (#5668)
* Update staverbs.dm

* Fixed oopsie

* UPDATE: Fixed jank

removed jank. interaction now takes longer, with the time greatly reduced by your COG stat. also adds a var to the computers to make it easier to track if they've been hacked, and notifies the user if the computer is hacked already when they try to interact with it.

* More staverb stuff

* Fixed excelsior implant jankiness and mines

* Changed faction check to antag_id check by request

* Last minor fix since today is the bulk merge day.

Co-authored-by: TorinoFermic <36856516+TorinoFermic@users.noreply.github.com>
Conflicts:
	code/game/objects/items/weapons/mine.dm