Minor blacklist update (#5698)
* Update staverbs.dm

* Fixed oopsie

* UPDATE: Fixed jank

removed jank. interaction now takes longer, with the time greatly reduced by your COG stat. also adds a var to the computers to make it easier to track if they've been hacked, and notifies the user if the computer is hacked already when they try to interact with it.

* More staverb stuff

* Fixed excelsior implant jankiness and mines

* Changed faction check to antag_id check by request

* Last minor fix since today is the bulk merge day.

* Update blacklist

* Revert "Update blacklist"

This reverts commit bdb660109a6b6e676193210953a982e6e63b0673.

* Properly updates blacklist (oops)

* Added more stuff that shouldn't spawn

* removed unnecessary vars

* Changed blacklist method by request

Co-authored-by: TorinoFermic <36856516+TorinoFermic@users.noreply.github.com>
Conflicts:
	code/modules/reagents/reagent_containers/glass/beaker.dm