Fix drag and drop to disposal
Adds checks to try to ensure the item being moved to the disposal is on a turf, rather than in an object. We do this because forceMove does not trigger any UI updates, and we want to retain moving things from turfs into disposal at some things can't be put in the disposal otherwise.