Bioreactor tweaks. (#1406)
* Technomancer removal 1

Removes multiple references in strings and comments to the Technomancers as an on-ship entity

* Technomancer removal 2: Engineering boogaloo

Renamed /datum/job/technomancer to /datum/job/engineer

* Adds verb to empty your hands

This verb allows mobs to drop what they're holding semi-forcibly. Rather than using unEquip, it skips past it and checks only for canremove to ensure the player does not drop implants like the arm blade.

Aside from being useful from an RP standpoint, the proc also is able to force bugged hands to drop what they're holding, and actually clears them.

* Fixes solar panel assembly assembly

Fixes the logic on the visible message from anchoring a solar assembly, and also moves the isturf check back to where it belongs so solar panel assemblies can be anchored and built.

* Revert "Technomancer removal 2: Engineering boogaloo"

This reverts commit a5c829dd9b80f1d19bbcb4d012d74ff11e566b37.

* Bioreactor console patch

Swaps out checking if MS is null, for just running check_MS. Also gives a TRUE value to MS already existing, so this can actually work.

Most of the time, this should be the same. But it will force the datum to be rebuilt if it's gone missing.

* Bioreactor stuck solution fix

Moves operational check to allow pumping solution out even if the bioreactor is broken. The only instance this should occur in is pipe blockage from contamination.

* Platform opacity fix

Opacity is a measure of how OPAQUE something is. Opaque is the opposite of transparent. Not sure how this ended up backwards.

* Changelog update

* Revert "Changelog update"

This reverts commit 941fcd837a66a03e2b129c44628c5b8e3b5ced50.