update jobs.dm
- fix #348
Set the surgical apron to cover arms to stop blood from spattering past the apron.

Not an ideal fix, as the coverage doesn't match the sprite now, but short of getting a brand new sprite, or refactoring how bloodsplatter is applied I don't really know how else we could fix this.