The Great Whiteboard Cleanup, Phase 1 (#1646)
* Hugs now give emergency sanity if your sanity is low. Plushies, too.

* Changes to Benzin leak code.

* More sanity checks in the hug/sanity code.

* Number of exoplanets to generate is now a config option.

* Rejuvinating agent is now less practical to make

* Gas masks now properly filter out chloramines.

* Ports ES13 species-specific chemistry reactions.

* Ghost sprites should no longer have PDAs.

* Adds backup SMES to the code, will need to map in later.

* Proper fix for dummymob email issue

* Initial dispatcher pass
Framework to allow dispatcher to call round ends and hivemind events.

* Minor tweaks to config wording.

* More config tweaks
* Adds in custom ship name option. Intended for use in upcoming dispatcher testing.
* Framework to allow dispatcher to announce an all-clear when hivemind is defeated and an announcement was made indicating its presence.

* Initial pass on simplemob bonus damage. Prolly refactor it later.

* Cleanup
I don't like how the simplemob bonus damage is implemented, probably going to refactor it later.

* Refactors simplemob bonus damage.

* Broad simplemob bonus damage settings pass
There'll probably be a few I missed.

* minor change to dispatcher reboot message

* Initial CA Maverick commit.

* I forgot you need to close out of the DME for it to #include the files

* Fixes issues preventing compile.

* NO CHANGES -- MESSAGE ONLY
This should be the last of the additions I need to do. Hopefully sometime soon I can start to test this shit. This will take a bit to test, since I'm only one guy who's working a 9-to-5 and burned the fuck out on Spessmans, so this will be tested when it gets tested.

* The batchfile compiler gives me trust issues.

* Vending machine changes.
Adds CA Maverick to vending machines, adds note at top of vending.dm asking other devs to copy it over to _eclipse.

* Increased price of Maverick (1500 -> 2000)

* Initial fixes for Maverick, fixes configuration error handling

* Simplemob bonus damage fixes.

* Fix for error handling.

* Tweaks to ticker and config. Removes debug code.

* Anti-powergaming measures, fixes EoR dispatch ping

* Dispatcher updates, fixes a few cock-ups in the sample config
Some idiot (me) got lazy and copy-pasted their active config...

* Final fix